### 2017-08-16 #364 @astrofrog
I'm on vacation for a couple of weeks, so I don't have time to review this at the moment.
