### 2017-08-20 e220491
pass origin in NFW potential
