### 2017-08-23 7541b1d
fix comparison with None
