+++ gala/potential/common.py
@@ -85,11 +85,11 @@ class CommonBase(object):
 
     # For comparison operations
     def __eq__(self, other):
+        if other is None or not hasattr(other, 'parameters'):
+            return False
+
         # the funkiness in the below is in case there are array parameters:
         par_bool = [(k1==k2) and np.all(self.parameters[k1] == other.parameters[k2])
                     for k1,k2 in zip(self.parameters.keys(), other.parameters.keys())]
         return np.all(par_bool) and (str(self) == str(other)) and (self.units == other.units)
-
-    def __ne__(self, other):
-        return not self.__eq__(other)
 
