### 2017-08-23 0fdfbff
added regression test
