### 2017-08-23 f982571
added a combine function for phasespacepositions and orbits
