### 2017-09-02 #364 @arfon
> I'd been looking at this code recently so I'd be happy to although I'll likely not get to it before next week. If that is okay, then I'd be happy to be a review for it.

⚡️ excellent. Thanks!
