### 2017-09-02 #388 @whedon
[REVIEW]: Gala: A Python package for galactic dynamics
**Submitting author:** @adrn (<a href="http://orcid.org/0000-0003-0872-7098">Adrian Price-Whelan</a>)
**Repository:** <a href="https://github.com/adrn/gala" target ="_blank">https://github.com/adrn/gala</a>
**Version:** v0.2.1
