### 2017-09-02 #364 @arfon
@crawfordsm see you over in #388 for the actual JOSS review.
