### 2017-09-07 4d98337
include bar angle in docs
