### 2017-09-08 fdd6e58
remove astropy monkey-patch
