+++ gala/__init__.py
@@ -15,21 +15,3 @@ if not _ASTROPY_SETUP_:
     from . import units
     from . import util
     from . import mpl_style
-
-    # TODO: remove with Astropy v1.3, which supports this
-
-    # Monkey-patch Quantity
-    import astropy.units as u
-
-    @classmethod
-    def _parse_quantity(cls, q):
-        try:
-            val,unit = q.split()
-        except:
-            val = q
-            unit = u.dimensionless_unscaled
-
-        return u.Quantity(float(val), unit)
-    u.Quantity.from_string = _parse_quantity
-
-    del u
