### 2017-09-09 7612920
fix docs import mpl_style [docs only]
