+++ gala/potential/potential/builtin/builtin_potentials.h
@@ -62,3 +62,4 @@ extern void logarithmic_gradient(double t, double *pars, double *q, int n_dim, d
 
 extern double longmuralibar_value(double t, double *pars, double *q, int n_dim);
 extern void longmuralibar_gradient(double t, double *pars, double *q, int n_dim, double *grad);
+extern double longmuralibar_density(double t, double *pars, double *q, int n_dim);
