+++ gala/potential/potential/builtin/builtin_potentials.c
@@ -950,3 +950,74 @@ void longmuralibar_gradient(double t, double *pars, double *q, int n_dim, double
     grad[1] = grad[1] + (gx*sin(pars[5]) + gy*cos(pars[5]));
     grad[2] = grad[2] + gz;
 }
+
+double longmuralibar_density(double t, double *pars, double *q, int n_dim) {
+    /*
+        Generated by sympy...
+
+        pars:
+        - G (Gravitational constant)
+        - m (mass scale)
+        - a
+        - b
+        - c
+        - alpha
+    */
+    double a = pars[2];
+    double b = pars[3];
+    double c = pars[4];
+
+    double x = q[0]*cos(pars[5]) + q[1]*sin(pars[5]);
+    double y = -q[0]*sin(pars[5]) + q[1]*cos(pars[5]);
+    double z = q[2];
+
+    double tmp0 = a - x;
+    double tmp1 = pow(tmp0, 2);
+    double tmp2 = pow(y, 2);
+    double tmp3 = pow(z, 2);
+    double tmp4 = pow(c, 2) + tmp3;
+    double tmp5 = sqrt(tmp4);
+    double tmp6 = b + tmp5;
+    double tmp7 = pow(tmp6, 2);
+    double tmp8 = tmp2 + tmp7;
+    double tmp9 = tmp1 + tmp8;
+    double tmp10 = sqrt(tmp9);
+    double tmp11 = -a + tmp10 + x;
+    double tmp12 = 1.0/tmp11;
+    double tmp13 = 1.0/tmp10;
+    double tmp14 = pow(tmp9, -1.5);
+    double tmp15 = 1.0/tmp4;
+    double tmp16 = tmp13*tmp3;
+    double tmp17 = tmp6/tmp5;
+    double tmp18 = pow(tmp4, -1.5);
+    double tmp19 = tmp15*tmp3*tmp7;
+    double tmp20 = 2*tmp2;
+    double tmp21 = a + x;
+    double tmp22 = pow(tmp21, 2);
+    double tmp23 = tmp22 + tmp8;
+    double tmp24 = sqrt(tmp23);
+    double tmp25 = 1.0/tmp24;
+    double tmp26 = tmp21 + tmp24;
+    double tmp27 = 1.0/tmp26;
+    double tmp28 = tmp25*tmp27;
+    double tmp29 = tmp11*tmp28;
+    double tmp30 = tmp11*tmp27/pow(tmp23, 1.5);
+    double tmp31 = 1.0/tmp23;
+    double tmp32 = pow(tmp26, -2);
+    double tmp33 = tmp11*tmp31*tmp32;
+    double tmp34 = tmp21*tmp25 + 1;
+    double tmp35 = tmp27*tmp34;
+    double tmp36 = tmp13*tmp15*tmp3*tmp7;
+    double tmp37 = -tmp13 + tmp29;
+    double tmp38 = tmp2*tmp37;
+    double tmp39 = tmp0*tmp13;
+    double tmp40 = tmp11*tmp27*tmp34 + tmp39 - 1;
+    return pars[1]/8.*tmp12*(2*tmp11*tmp32*pow(tmp34, 2) +
+        tmp12*tmp13*tmp38 + tmp12*tmp36*tmp37 + tmp12*tmp40*(-tmp39 + 1) +
+        tmp13*tmp17 - tmp13*tmp20*tmp25*tmp27 + tmp13*(-tmp1/tmp9 + 1) + tmp13 -
[…]
