### 2017-09-10 066724e
ignore warnings - needed for new pytest
