+++ setup.cfg
@@ -7,10 +7,15 @@ all_files = 1
 upload-dir = docs/_build/html
 show-response = 1
 
-[pytest]
+[tool:pytest]
 minversion = 2.2
 norecursedirs = build docs/_build
 doctest_plus = enabled
+filterwarnings = 
+    error
+    ignore::ImportWarning
+    ignore::UserWarning
+    ignore::RuntimeWarning
 
 [ah_bootstrap]
 auto_use = True
