### 2017-09-10 adb18c3
expect hamiltonian not potential
