+++ gala/dynamics/tests/test_util.py
@@ -38,10 +38,13 @@ def test_estimate_dt_n_steps():
     nperiods = 128
     pot = NFWPotential.from_circular_velocity(v_c=1., r_s=10., units=galactic)
     w0 = [10.,0.,0.,0.,0.9,0.]
-    dt,n_steps = estimate_dt_n_steps(w0, pot, n_periods=nperiods, n_steps_per_period=256,
+
+    H = Hamiltonian(pot)
+    dt,n_steps = estimate_dt_n_steps(w0, H, n_periods=nperiods,
+                                     n_steps_per_period=256,
                                      func=np.nanmin)
 
-    orbit = Hamiltonian(pot).integrate_orbit(w0, dt=dt, n_steps=n_steps)
+    orbit = H.integrate_orbit(w0, dt=dt, n_steps=n_steps)
     T = orbit.estimate_period()
     assert int(round((orbit.t.max()/T).decompose().value)) == nperiods
 
