### 2017-09-10 bd7453d
allow passing integrate kwargs
