### 2017-09-10 4c15a58
aligned orbit should have same shape
