+++ setup.cfg
@@ -16,6 +16,7 @@ filterwarnings =
     ignore::ImportWarning
     ignore::UserWarning
     ignore::RuntimeWarning
+    ignore::DeprecationWarning
 
 [ah_bootstrap]
 auto_use = True
