+++ setup.cfg
@@ -11,12 +11,11 @@ show-response = 1
 minversion = 2.2
 norecursedirs = build docs/_build
 doctest_plus = enabled
-filterwarnings = 
-    error
+# For the warnings filter, see: https://github.com/astropy/astropy/issues/6025
+filterwarnings =
     ignore::ImportWarning
     ignore::UserWarning
     ignore::RuntimeWarning
-    ignore::DeprecationWarning
 
 [ah_bootstrap]
 auto_use = True
