### 2017-09-11 4f00607
fix warnings
