### 2017-09-11 e5c54e6
clean up action angle docs and fix marker=None in plot calls
