### 2017-09-13 #388 @arfon
> Sorry about that -- am I suppose to be able to check or edit the comment at the top?

Yep, you *should* be able to edit the checklist at the top of the issue.
