### 2017-09-13 #388 @arfon
> Ah that's not the case. Unfortunately, I can't edit the top comment. Happy to copy and past and use another comment though.

Sure, that works too.
