### 2017-09-13 #388 @crawfordsm
Conflict of interest

Code of Conduct

- []   I confirm that I read and will adhere to the JOSS code of conduct.

General checks

- []   Repository: Is the source code for this software available at the repository url?
- []   License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?
- []   Version: Does the release version given match the GitHub release (v0.2.1)?
- []   Authorship: Has the submitting author (@adrn) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

- []   Automated tests: Are there automated tests or manual steps described so that the function of the software can be verified?

 - []  Authors: Does the paper.md file include a list of authors with their affiliations?
- []   References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
