### 2017-09-13 #388 @crawfordsm
Using this as notepad for high level comments -- will transfer to the repository once finished with review

The gala software package provides tools for galactic dynamic measurements.  It is python package that wraps lower level, higher performance, and provides a user friendly interface to those lower level codes.  It is well integrated with astropy, and provides documentation of its functionality.   The code is generally useful to the astronomical community in both terms of research and education.

* Contributors are listed in the Zenodo DOI and on the README with how to cite.  Primary author is listed in paper.md, but not contributing authors.   Might be useful to have a AUTHORS.md in the main repository

* The one weakness of the documentation is really a high level justification of the project.  I found this information in the paper.md and I would suggest including this material in the main documentation.

Minor comments:
* 'why' page might be better titled 'Benefits of Gala'
