### 2017-09-14 03e6dfa
added an AUTHORS file with contributors
