### 2017-09-14 4b1809a
add context to docs page. address comments in openjournals/joss-reviews#388
