### 2017-09-14 #388 @adrn
Thanks for the comments! I addressed the ones you left so far:
* I added an [AUTHORS.rst](https://github.com/adrn/gala/blob/master/AUTHORS.rst) file, and now include it in the documentation
* I added the high-level justification / context to the main page of the documentation
* I renamed the 'why' page
