### 2017-09-18 ca9b67b
add density function for Kepler
