### 2017-09-18 53199a0
forgot to add kepler C func to classes
