### 2017-09-18 eba61d2
fix mass enclosed for negative mass
