### 2017-09-19 #388 @crawfordsm
Comments above have been address.

The functionality and performance have been tested as well and they match the claims that have been reported.  Overall, the package is well documented with good examples allowing example problems to be reported.

Final comments to be address:
* It would be useful to add testing information to the installation page

*  it would be good to indicate whether contributions are welcome and the best way to make them.   Seeking support and asking questions is clearly indicated but how to make contributions is not.   Perhaps a contributing section in the README would help.

Suggestion for the future:   It might be good to include a gallery page including further examples.
