### 2017-09-23 26f9905
added stub testing doc [ci skip]
