+++ docs/testing.rst
@@ -1,0 +1,12 @@
+.. _gala-test:
+
+=================
+Running the tests
+=================
+
+The tests are written assuming they will be run with `pytest <>`_ using the
+Astropy custom test runner. To run the full set of tests, we recommend
+installing (in addition to the main dependencies listed in
+:ref:`gala-install`)::
+
+    cython scipy matplotlib pyyaml h5py sympy qt
