### 2017-09-23 e405cbc
add document describing how to run the tests and add deve environment file [docs only]
