### 2017-09-23 4464908
added a how to contribute section of the docs [docs only]
