### 2017-09-23 #388 @adrn
Good ideas!

http://gala-astro.readthedocs.io/en/latest/

- I added a link and page "Running the tests" linked from the main docs
- I added a section "How to contribute" to the main docs page

I like the idea of having a gallery for examples / tutorials - I'll explore that when I've written some more content.
