### 2017-09-24 #388 @crawfordsm
Looks great to me!   I noticed the other authors aren't listed in the paper.md file.  Last thing needed to updated for the checklist to be complete.
