### 2017-09-24 #388 @adrn
That was actually intentional - is there a requirement that all contributors are authors?
