### 2017-09-24 #388 @adrn
(If not, I would prefer to keep it as is)
