### 2017-09-24 #388 @crawfordsm
Seems like a question for @arfon, but JOSS does seem to ask to include the list of Authors of the software in the authors.

If the JOSS entry will be the main citation for the software, then I'd suggest including them as you do currently include them under 'how to cite' section.   I could understand having other justifications and it would be good to hear them though this is likely up to @arfon
