### 2017-09-25 #388 @arfon
> That was actually intentional - is there a requirement that all contributors are authors?

We ask reviewers to check that the list of authors is 'correct' but this is ultimately somewhat subjective based on what seems reasonable. See [this thread for a similar discussion](https://github.com/openjournals/joss-reviews/issues/348#issuecomment-325604616)

@adrn - I'm assuming you're not including the other three contributors as they all made relatively minor contributions to the package? If so, perhaps you could acknowledge their contributions in the README or somewhere in the body of the paper?
