### 2017-09-25 4b3f67f
add explicit link to authors [ci skip]
