### 2017-09-25 #388 @adrn
> I'm assuming you're not including the other three contributors as they all made relatively minor contributions to the package?

Exactly

>  If so, perhaps you could acknowledge their contributions in the README or somewhere in the body of the paper?

Done - I added a link to the AUTHORS.rst file to the README, and added an acknowledgements section to the paper.
