### 2017-09-25 #388 @arfon
Thanks @adrn. @crawfordsm - can you confirm we're good to accept here?
