### 2017-09-28 #388 @crawfordsm
Thanks for adding it in the acknowledgements.   This is obviously a difficult question and one with out a very clearly defined answer.     A statement from the JOSS board might be very helpful going ahead in the future also to give some ideas, and leadership, for the community as to what are the best suggested and most fair practices.  I know I have had difficulty in the past in terms of balancing wanting to reflect what is >90% my own work along with the relatively small contributions of others and it is a difficult thing to balance.

The paper format is  certainly outdated to properly attribute the work of different contributors as authors.   As a suggestion to @arfon, I might suggest  two parts to the attribution section of the paper.md file, which would be Authors and Contributors.  There might be some overlap between the two, but authors would be defined as those that have made significant scientific/programming contribution to the code and contributors are people that have made more programmatic contributions to the code.    The option or suggest to include a description of the contribution of each author might be helpful as well.

For @adrn , I think my concern is how are you going to suggest the paper is cited?  Publications and citations are the credit of the academic world, and I like that your currently suggested manner of citations is the zenodo link that includes your contributors.    If you were going to change this to being _only_ the JOSS paper, I think I'd find that problematic as most of your contributors are academics.  On the other hand, the contributions of the others are relatively minor to your overall development of the package.  Personally, I err on the side of included more people, but I definitely understand arguments on the other side and could see the acknowledgement as being sufficient based on the level of contribution.

Nonetheless, as there is no definitive policy and this work is predominately yours, I'd default to what you thought best reflected the package and contribution of others.
