### 2017-09-28 adrn/gala#100 @adrn
Authorship on JOSS paper
@bsipocz @smoh @syrte

I've submitted a paper to [JOSS](joss.theoj.org) describing Gala. It's [currently under review](https://github.com/openjournals/joss-reviews/issues/388), and you can read the review and response at the link. Please let me know by 1 Oct if you'd like to be listed as co-authors on the paper, in which case please send me your current affiliation and ORCID.
