### 2017-09-28 #388 @crawfordsm
Accept

On Thu, 28 Sep 2017 at 10:28 AM Adrian Price-Whelan <
notifications@github.com> wrote:

> I agree that this is a tough and confusing topic. For lack of a clear set
> of rules, I'll just default to following the same rules I use for other
> scientific papers: in my own arbitrary system (adapted from @davidwhogg
> <https://github.com/davidwhogg>), if someone contributes a "minor change"
> (yes, this is subjective), they are added to the acknowledgements and
> explicitly asked if they would like to be included as authors of the paper.
> If they say yes, they are expected to contribute more significantly to the
> paper. I do think there should be a threshold to receiving citation-based
> attribution for contributions in general.
>
> Since there is no obvious analogy to the above in this case (the "paper"
> is not the main work effort; the codebase is), I'll ask if the contributors
> would like to be authors on the paper, and I'll be more lenient about
> expecting further contributions.
>
