### 2017-09-28 adrn/gala#100 @bsipocz
Hey @adrn! I can only repeat what I said for other papers in the past, namely while getting on papers is always exciting and a great feedback, I feel this time I didn't do enough with the mere 8 commits. It's just not fair towards others (this time you), who did the actual coding and serious bugfixing.
