### 2017-09-28 #388 @crawfordsm
Sorry for my premature acceptance, but was going to say

Accept with Adrian's final decision about author list.  I do not need to
review again unless anything arfon has any further requests.

On Thu, 28 Sep 2017 at 11:08 AM Adrian Price-Whelan <
notifications@github.com> wrote:

> (I gave them until 1 Oct to respond!)
>
