### 2017-10-01 #388 @adrn
I've heard from 2/3 committers (see: https://github.com/adrn/gala/issues/100) - keep the author list as is. Sounds like this is ready to go then @arfon?
