+++ setup.py
@@ -47,7 +47,7 @@ LONG_DESCRIPTION = package.__doc__
 builtins._ASTROPY_PACKAGE_NAME_ = PACKAGENAME
 
 # VERSION should be PEP386 compatible (http://www.python.org/dev/peps/pep-0386)
-VERSION = '0.3.dev'
+VERSION = '0.2.2'
 
 # Indicates if this version is a release version
 RELEASE = 'dev' not in VERSION
