+++ CHANGES.md
@@ -1,3 +1,12 @@
+0.3 (unreleased)
+================
+
+New Features
+------------
+
+API-breaking changes
+--------------------
+
 0.2.2 (2017-10-07)
 ==================
 
