### 2017-10-07 #388 @adrn
DOI is still catching up, but it should be here soon:
https://doi.org/10.5281/zenodo.1004642
